Update 05_other.yml
Mi Pad by default has 2 /syetm partition, each 640Mb. It is not enogh to install LineageOS and Gapps.  So have to merge 2 system partition in one 1.2Gb system partition.To dod this follow instruction http://en.miui.com/thread-118294-1-1.html.That's the way it was after the merger.
Please add it,Thank you
[My English is not very good, please forgive me]